requirements.txt
# ── Core AI / Agent framework ─────────────────────────────────────────────────
crewai>=0.70.0
crewai-tools>=0.8.0

# ── LLM providers ─────────────────────────────────────────────────────────────
openai>=1.30.0
groq>=0.9.0

# ── RAG pipeline ─────────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# ── Web search ────────────────────────────────────────────────────────────────
duckduckgo-search>=6.2.0

# ── API server ────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
httpx>=0.27.0
